% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor-pay.R
\name{calculate_instructor_pay}
\alias{calculate_instructor_pay}
\title{Calculate instructor pay per person per class type}
\usage{
calculate_instructor_pay(
  hourly_rate,
  class_type,
  number_of_courses,
  instructor
)
}
\arguments{
\item{hourly_rate}{\itemize{
\item hourly pay rate
}}

\item{class_type}{\itemize{
\item class type. One of "member" or "public".
}}

\item{number_of_courses}{\itemize{
\item number of courses taught by a given instructor
}}

\item{instructor}{\itemize{
\item name of instructor
}}
}
\value{
numeric
}
\description{
Calculate instructor pay per person per class type
}
