% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_calendar_event}
\alias{create_calendar_event}
\title{Subject (Required) The name of the event
Start Date (Required) The first day of the event
Example: 05 / 30 / 2020
Start Time - The time the event begins
Example: 10:00 AM
End Date The last day of the event
Example: 05 / 30 / 2020
End Time}
\usage{
create_calendar_event(class_schedule)
}
\description{
Subject (Required) The name of the event
Start Date (Required) The first day of the event
Example: 05 / 30 / 2020
Start Time - The time the event begins
Example: 10:00 AM
End Date The last day of the event
Example: 05 / 30 / 2020
End Time
}
