% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM.R
\name{DFO-operating-models}
\alias{DFO-operating-models}
\alias{DFO_DEMO_decreaseM}
\alias{DFO_DEMO_increaseM}
\alias{DFO_4X5Y_Haddock_2015}
\alias{DFO_Arrowtooth_Flounder_2014}
\alias{DFO_BoF_Herring_2018}
\alias{DFO_EGB_Haddock_2019}
\alias{DFO_GSL_Atlantic_Cod_2018}
\alias{DFO_GSL_SS_Herring_2019}
\alias{DFO_HG_Herring_2019}
\alias{DFO_Inside_QB_Rockfish_2020}
\alias{DFO_Inside_YE_Rockfish_2019}
\alias{DFO_Pacific_Cod_2020}
\alias{DFO_Pacific_Hake_2019}
\alias{DFO_POP_2016}
\alias{DFO_SoG_Herring_2019}
\alias{DFO_WCVI_Herring_2019}
\alias{Peruvian_Anchoveta_2008}
\alias{OM}
\title{Pre-built operating models}
\format{
An object of class \linkS4class{OM}.
}
\description{
A set of operating models built into RPC.
}
\section{Operating models for Canadian fisheries}{

\itemize{
\item \code{DFO_DEMO_decreaseM}: Example OM with decreasing M over time
\item \code{DFO_DEMO_increaseM}: Example OM with increasing M over time
\item \code{DFO_4X5Y_Haddock_2015}: 2015 VPA assessment imported by \link[MSEtool]{VPA2OM}
\item \code{DFO_Arrowtooth_Flounder_2014}: 2014 assessment, MPD estimates imported by \link[MSEtool]{iSCAM2OM}
\item \code{DFO_BoF_Herring_2018}: operating model conditioned by \link[SAMtool]{RCM}
\item \code{DFO_EGB_Haddock_2019}: 2022 assessment (data up to 2019) imported by \link[MSEtool]{WHAM2OM}.
\item \code{DFO_GSL_Atlantic_Cod_2018}: 2018 assessment, imported by \link[MSEtool]{Assess2OM}.
\item \code{DFO_GSL_SS_Herring_2019}: SCA assessment of southern Gulf of St. Lawrence Atlantic Herring, Spring Spawning Component, imported by \link[MSEtool]{Assess2OM}
\item \code{DFO_HG_Herring_2019}: 2019 iSCAM assessment update, MPD estimates imported by \link[MSEtool]{iSCAM2OM}
\item \code{DFO_Inside_QB_Rockfish_2020}: Preliminary inside quillback rockfish operating model, conditioned by \link[SAMtool]{RCM}
\item \code{DFO_Inside_YE_Rockfish_2019}: Base 2019 inside yelloweye rockfish operating model, conditioned by \link[SAMtool]{RCM}
\item \code{DFO_Pacific_Cod_2020}: Base 2020 iSCAM assessment update, MCMC output imported by \link[MSEtool]{iSCAM2OM}
\item \code{DFO_Pacific_Hake_2019}: Base 2019 SS3 assessment, output imported by \link[MSEtool]{SS2OM}
\item \code{DFO_POP_2016}: Assessment with MCMC output imported by \link[MSEtool]{Awatea2OM}
\item \code{DFO_SoG_Herring_2019}: 2019 iSCAM assessment update, MPD estimates imported by \link[MSEtool]{iSCAM2OM}
\item \code{DFO_WCVI_Herring_2019}: 2019 iSCAM assessment update, MPD estimates imported by \link[MSEtool]{iSCAM2OM}
}
}

\section{Other operating models}{

\itemize{
\item \code{Peruvian_Anchoveta_2008}
}
}

