% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_class_schedule}
\alias{get_class_schedule}
\title{get_class_schedule}
\usage{
get_class_schedule(instructor = NULL, sheet)
}
\arguments{
\item{instructor}{(chr) name of instructor}

\item{sheet}{(str) name of sheet to use}
}
\value{
tibble
}
\description{
get_class_schedule
}
\examples{
\dontrun{get_class_schedule("First Last", sheet = "2026 Winter")}
}
