% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_exclusions}
\alias{get_exclusions}
\title{get exclusion dates}
\usage{
get_exclusions(start_date, sessions, class_frequency)
}
\arguments{
\item{start_date}{(date) start date}

\item{sessions}{(integer) number of sessions}

\item{class_frequency}{(bool) if the classes happen class_frequency}
}
\description{
get exclusion dates
}
