% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_pick_up_date}
\alias{get_pick_up_date}
\title{Public class pick up date
Sets the pick up date 2 weeks after the last public class}
\usage{
get_pick_up_date(last_class)
}
\arguments{
\item{last_class}{the date of the last class}
}
\value{
date
}
\description{
Public class pick up date
Sets the pick up date 2 weeks after the last public class
}
