% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{hist_resample_recruitment}
\alias{hist_resample_recruitment}
\title{Resample recruitment deviations}
\usage{
hist_resample_recruitment(
  x,
  dist = c("Lognormal", "Pareto"),
  mu = 1,
  LnSD = 0.7,
  LnAC = 0,
  Pshape = 1.1,
  figure = TRUE,
  nsim_plot = 5
)
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{dist}{Character to denote to sample either from a lognormal distribution or Pareto distribution.}

\item{mu}{The mean of the distribution (default = 1).}

\item{LnSD}{If Lognormal, the standard deviation.}

\item{LnAC}{If Lognormal, the autocorrelation (in log-space).}

\item{Pshape}{If Pareto, the shape parameter. See \link[EnvStats]{Pareto}. The location parameter is calculated such that the mean = 1.}

\item{figure}{Whether to plot the sampled deviations.}

\item{nsim_plot}{The number of simulations to plot if figure is TRUE.}
}
\value{
A list with an updated matrix for \code{OM@cpars$Perr_y}.
}
\description{
A function to generate new recruitment deviations (with mean = 1) for the projection period.
}
