% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{make-MP}
\alias{make-MP}
\alias{make_FixC_MP}
\alias{make_FixF_MP}
\alias{make_RPC_MP}
\alias{make_HCR_name}
\title{Functions for making management procedures}
\usage{
make_FixC_MP(ratio = 1)

make_FixF_MP(ratio)

make_RPC_MP(input, verbose = TRUE)

make_HCR_name(input, type = c("label", "description"))
}
\arguments{
\item{ratio}{Numeric, the ratio of projected F or catch relative to that in the last historical year.}

\item{input}{The input list from a shiny server session.}

\item{verbose}{Whether to return information on the MP in the server log.}

\item{type}{Whether a label or full description is returned.}
}
\description{
Various functions that generate MPs
}
\details{
\code{make_FixC_MP} generates a fixed catch MP.

\code{make_FixF_MP} generates a fixed F MP. Ensure that \code{OM@qinc <- OM@qcv <- c(0, 0)}.

\code{make_RPC_MP} generates MP using a harvest control rule in the App.

\code{make_HCR_name} returns a default name or description of the harvest control rule.
}
