% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proj.R
\name{make_PMobj}
\alias{make_PMobj}
\title{Make a performance metric}
\usage{
make_PMobj(
  x,
  type = c("SSB", "SSB0", "SSBMSY", "F", "SPR", "Catch"),
  frac = 0.4,
  year_range,
  label,
  ...
)
}
\arguments{
\item{x}{An object of class \linkS4class{MSE}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEproj} which is
the MSE object.}

\item{type}{The response variable for the performance metric.}

\item{frac}{The reference value for calculating probability of exceeding this value.}

\item{year_range}{The length two vector for the range of years (relative to OM@CurrentYr)
during which to calculate the performance measure.}

\item{label}{Optional character that describes the performance measure.}

\item{...}{Additional arguments depending on \code{type}. See details.}
}
\value{
A \linkS4class{PMobj} object.
}
\description{
Function that defines and returns performance measures.
}
\details{
\code{type = "SSB"} calculates SSB relative to historical SSB. Provide the reference year with argument \code{SSBhist_yr}.

\code{type = "SSB0"} calculates SSB relative to SSB0 (either "Dynamic", "Initial", or "Equilibrium").
Provide the type with argument \code{SSB0_type}.

\code{type = "SSBMSY"} calculates SSB relative to SSBMSY.

\code{type = "F"} calculates F relative to FMSY.

\code{type = "SPR"} calculates SPR relative to frac.

\code{type = "SSBMSY"} calculates SSB relative to historical SSB. Provide the reference year with argument \code{Chist_yr}.
}
\examples{
\donttest{
library(MSEtool)
MSE <- MSEtool::runMSE(OM = MSEtool::testOM, MPs = c("NFref", "FMSYref"))
PM <- make_PMobj(MSE, type = "SSB", frac = 0.75, year_range = 51:100, SSBhist_yr = 50)
}
}
