% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pick_up_date}
\alias{pick_up_date}
\title{pick_up_date
Sets the pick up date 2 weeks after the last public class}
\usage{
pick_up_date(last_class)
}
\arguments{
\item{last_class}{the date of the last class}
}
\value{
date
}
\description{
pick_up_date
Sets the pick up date 2 weeks after the last public class
}
