% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res_hist.R, R/plot_res_hist_diagnostic.R
\name{plot-Hist-diagnostic}
\alias{plot-Hist-diagnostic}
\alias{hist_Rmax}
\alias{hist_RpS90}
\title{Biomass limit reference points methods}
\usage{
hist_Rmax(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))

hist_RpS90(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{figure}{Logical, whether to return a stock-recruit figure (TRUE) or data frame reporting the limit reference point (FALSE).}

\item{prob_ratio}{If NA, returns the limit reference point. Otherwise, a numeric that indicates the threshold of exceeding the reference point.
Functions return annual probabilities of exceeding this threshold.}

\item{prob_ylim}{The y-axis range of the figure if prob_ratio is a numeric. Only used if \code{figure = TRUE}.}
}
\description{
Two figures showing diagnostics associated with methods for identifying biomass limit reference points using
the stock-recruit relationship.
}
\details{
\code{hist_Rmax} returns either \code{SSB50\%Rmax}, the SSB corresponding to 50% of maximum recruitment from the stock-recruit relationship
(with regression lines of recruits vs. spawners above and below this value), or the annual probability of exceeding some ratio of
\code{SSB50\%Rmax}.

\code{hist_RpS90} returns either \code{SSB 90\%ile R/S}, the SSB corresponding to the intersection of the 90the percentile of both historical recruitment
and recruits-per-spawner, or the annual probability of exceeding some ratio of \code{SSB 90\%ile R/S}.
}
\examples{
Hist <- MSEtool::runMSE(Hist = TRUE)

hist_Rmax(Hist)
hist_Rmax(Hist, prob_ratio = 0.5)

hist_Rmax(Hist, figure = FALSE)
hist_Rmax(Hist, figure = FALSE, prob_ratio = 0.5)
}
\references{
Mace, P.M. 1994. Relationships between Common Biological Reference Points Used as Thresholds and Targets of Fisheries Management Strategies.
CJFAS. 51:110-122. https://doi.org/10.1139/f94-013

Myers, et al. 1994. In search of thresholds for recruitment overfishing. ICES JMS. 51:191–205. https://doi.org/10.1006/jmsc.1994.1020
}
