% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res_hist.R
\name{plot-Hist-prob}
\alias{plot-Hist-prob}
\alias{hist_SSB}
\alias{hist_SSBMSY}
\alias{hist_SSB0}
\alias{hist_BvsSP}
\alias{hist_R}
\alias{hist_RpS}
\alias{hist_SPR}
\alias{hist_exp}
\alias{hist_Fmed}
\title{Plot historical dynamics with probabilities}
\usage{
hist_SSB(x, figure = TRUE, SSB_y = NA, prob_ratio = NA, prob_ylim = c(0, 1))

hist_SSBMSY(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1.5))

hist_SSB0(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))

hist_BvsSP(x, figure = TRUE)

hist_R(
  x,
  figure = TRUE,
  SR_only = FALSE,
  SR_xlim,
  SR_ylim,
  SR_y_RPS0,
  SR_include = 1:3
)

hist_RpS(x, figure = TRUE)

hist_SPR(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))

hist_exp(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))

hist_Fmed(x, figure = TRUE, prob_ratio = NA, prob_ylim = c(0, 1))
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{figure}{Logical, whether to return a figure (TRUE) or data frame (FALSE).}

\item{SSB_y}{The year (relative to OM@CurentYr) to compare annual SSB values (only if prob_ratio is a numeric).}

\item{prob_ratio}{If NA, returns annual values. Otherwise, a numeric that indicates a threshold. Functions return
annual probabilities of exceeding this threshold.}

\item{prob_ylim}{The y-axis range of the figure if prob_ratio is a numeric. Only used if \code{figure = TRUE}.}

\item{SR_only}{Logical, whether to plot annual recruitment (FALSE) or the stock-recruit figure (TRUE). Only used if \code{figure = TRUE}.}

\item{SR_xlim}{Optional x-axis range for the stock recruit plot. Only used if \code{SR_only = TRUE} and \code{figure = TRUE}.}

\item{SR_ylim}{Optional y-axis range for the stock recruit plot. Only used if \code{SR_only = TRUE} and \code{figure = TRUE}.}

\item{SR_y_RPS0}{The year (relative to OM@CurrentYr) for which to plot unfished recruits per spawner,
Only used if \code{SR_only = TRUE} and \code{figure = TRUE}, and \code{any(SR_include == 3)}.}

\item{SR_include}{A vector including any of c(1, 2, 3) that indicates what to plot in the stock-recruit figure. 1 = individual S-R pairs,
2 = stock-recruit relationship, 3 = reference recruits-per-spawner (R/S) lines
(maximum R/S corresponding to stock-recruit alpha, median historical R/S, and year-specific unfished R/S).}
}
\value{
If \code{figure = TRUE}, various plots using either base graphics or ggplot2 showing annual values (\code{prob_ratio = NA}) or
probabilities (\code{prob_ratio} is numeric). If \code{figure = FALSE}, returns a data frame displaying quantiles (\code{prob_ratio = NA})
or probabilities (\code{prob_ratio} is numeric).
}
\description{
Generate figures associated with probabilities
}
\details{
\code{hist_SSB} returns annual SSB or the annual probability that SSB exceeds some historical value (corresponding to the year in
argument \code{SSB_y}).

\code{hist_SSBMSY} returns annual SSB/SSBMSY or the annual probability that SSB/SSBMSY exceeds \code{prob_ratio}.

\code{hist_SSB0} returns annual SSB/SSB0 or the annual probability that SSB/SSB0 exceeds \code{prob_ratio}.

\code{hist_BvsSP} returns annual surplus production (annual change in biomass - catch) and per capita surplus production.

\code{hist_R} returns either annual recruitment (as a figure or table) or a stock recruit figure.

\code{hist_RpS} returns annual recruits-per-spawner.

\code{hist_SPR} returns annual spawning potential ratio.

\code{hist_exp} returns annual F and F/FMSY.

\code{hist_Fmed} returns F and Fmed (the fishing mortality corresponding to the historical median recruits per spawner,
frequently also referred to as the F-replacement F).
}
\examples{
Hist <- MSEtool::runMSE(Hist = TRUE)

hist_SSB(Hist)
hist_SSB(Hist, SSB_y = 50, prob_ratio = 0.5)

hist_SSB(Hist, figure = FALSE)
hist_SSB(Hist, figure = FALSE, SSB_y = 50, prob_ratio = 0.5)
}
\author{
Q. Huynh
}
