% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRplot.R
\name{plot-MP}
\alias{plot-MP}
\alias{HCR_plot}
\alias{CurF_plot}
\alias{CurC_plot}
\title{Functions to plot MPs}
\usage{
HCR_plot(input)

CurF_plot(x, F_ratio = 1)

CurC_plot(x, C_ratio = 1)
}
\arguments{
\item{input}{The input list from the shiny server}

\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{F_ratio}{The ratio of projection F and that from the last historical year.}

\item{C_ratio}{The ratio of projection catch and that from the last historical year.}
}
\description{
Various functions to plot the MP on the RPC App.
}
\details{
\code{HCR_plot} plots the selected harvest control rule.

\code{CurF_plot} plots the projection F relative to historical F.

\code{CurC_plot} plots the projection catch relative to historical.
}
