% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proj.R
\name{plot-MSE}
\alias{plot-MSE}
\alias{proj_plot}
\alias{prob_plot}
\alias{stoch_plot}
\alias{hist_sim}
\alias{lollipop_plot}
\alias{tradeoff_plot}
\alias{radar_plot}
\title{Plot output from MSE}
\usage{
proj_plot(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Catch"))

prob_plot(x, PM_list = list(), xlim = NULL, ylim = NULL, figure = TRUE)

stoch_plot(
  x,
  MPstoch,
  qval = 0.9,
  type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Catch")
)

hist_sim(
  x,
  MSEhist,
  MP,
  sims,
  type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Catch")
)

lollipop_plot(x, PM_list)

tradeoff_plot(x, PMx, PMy, xlab, ylab)

radar_plot(x, PM_list, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{MSE}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEproj} which is
the MSE object.}

\item{MSEhist}{If \code{x} is an MSE object, then a \linkS4class{Hist} object may also be needed.}

\item{type}{Character describing the type of output.}

\item{PM_list}{A list of \linkS4class{PMobj} objects, preferably, generated by \link{make_PMobj}.}

\item{xlim}{The year range over which to plot annual probabilities of the performance measure if \code{figure = TRUE}.}

\item{ylim}{The y-axis range over which to plot annual probabilities of the performance measure if \code{figure = TRUE}.}

\item{figure}{Logical, whether to plot the performance measure (TRUE) or return a matrix (FALSE).}

\item{MPstoch}{A character vector of MPs to plot.}

\item{qval}{The quantile of the confidence interval to plot.}

\item{MP}{Character, name of MP to plot.}

\item{sims}{Numeric vector (up to length three) indicating the simulations to plot.}

\item{PMx}{A \linkS4class{PMobj} object to show on the x-axis.}

\item{PMy}{A \linkS4class{PMobj} object to show on the y-axis.}

\item{xlab}{Name of the performance metric on the x-axis.}

\item{ylab}{Name of the performance metric on the y-axis.}

\item{...}{Additional arguments to \link[ggspider]{spider_web}.}
}
\value{
Various plots using base graphics or ggplot2.
}
\description{
Figures of results of the closed-loop simulation by MP
}
\details{
\code{proj_plot} generates median trajectories (historical and projected) with reference points if applicable.

\code{prob_plot} generates either a figure or a table of performance metrics.

\code{stoch_plot} generates median trajectories with confidence intervals.

\code{hist_sim} plots individual simulations (up to three).

\code{lollipop_plot} plots all performance measures for all MPs.

\code{tradeoff_plot} plots a comparison of two performance measures for all MPs.

\code{radar_plot} plots all performance measures for all MPs, where each metric is plotted along the vertex of a polygon.
}
\examples{
library(DLMtool)
Hist <- MSEtool::runMSE(Hist = TRUE)
MSE <- MSEtool::Project(
  Hist,
  MPs = c("NFref", "AvC"),
  extended = TRUE # Needed for type = "SP"
)
MSE@Hist <- Hist
proj_plot(MSE, Hist, type = "SP")

}
\author{
Q. Huynh
}
