% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\alias{hist_bio}
\alias{hist_future_recruit}
\alias{hist_bio_schedule}
\alias{hist_bio_change}
\alias{hist_growth_I}
\alias{hist_growth_II}
\alias{hist_spatial}
\alias{hist_sel}
\alias{hist_YieldCurve}
\alias{hist_resample_recruitment}
\alias{hist_SRR_change}
\alias{hist_phi0}
\alias{hist_per_recruit}
\title{Plot historical dynamics of the operating model. See details below.}
\usage{
hist_bio(x, figure = TRUE)

hist_future_recruit(x, figure = TRUE)

hist_bio_schedule(
  x,
  var = c("Len_age", "Wt_age", "Mat_age", "M_ageArray"),
  n_age_plot,
  yr_plot,
  sim,
  figure = TRUE
)

hist_bio_change(
  x,
  var = c("Wt_age", "M_ageArray"),
  change_mean = 0,
  change_slope = 0,
  n_age_plot = 10,
  sim = 1,
  figure = TRUE
)

hist_growth_I(x)

hist_growth_II(x)

hist_spatial(x, type = c("par", "matrix", "all"), ...)

hist_sel(x, yr, maturity = TRUE, figure = TRUE)

hist_YieldCurve(x, yr_bio, yr_sel, F_range, figure = TRUE)

hist_resample_recruitment(
  x,
  dist = c("Lognormal", "Pareto"),
  mu = 1,
  LnSD = 0.7,
  LnAC = 0,
  Pshape = 1.1,
  figure = TRUE,
  nsim_plot = 5
)

hist_SRR_change(x, SR_new = 1, h_mult = 1, y_fit, figure = TRUE)

hist_phi0(x, figure = TRUE)

hist_per_recruit(x, yr_bio, yr_sel, F_range, figure = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{figure}{Logical, whether to return a figure (TRUE) or the updated array (FALSE) for \code{hist_bio_change}.}

\item{var}{A string to indicate which object to plot from OM@cpars.}

\item{n_age_plot}{The number of ages to plot in the left figure.}

\item{yr_plot}{The year (relative to OM@CurrentYr) to plot for the right figure.}

\item{sim}{The simulation to plot for the left figure.}

\item{change_mean}{The percent change in the mean of the new parameters relative to the old.}

\item{change_slope}{The percent change year-over-year in the projection parameters relative to the last projection year.}

\item{type}{Type of spatial plot}

\item{...}{arguments to \link[MSEtool]{plot_mov}}

\item{yr}{A length-2 vector for the years (relative to OM@CurrentYr) to plot selectivity.}

\item{maturity}{Logical, whether to plot maturity along with selectivity.}

\item{yr_bio}{The year (relative to OM@CurrentYr) for the biological parameters (growth, M, maturity).}

\item{yr_sel}{The year (relative to OM@CurrentYr) for the selectivity parameters.}

\item{F_range}{Length two vector for the range of F to plot the yield curve. By default, \code{c(1e-8, 3 * max(M))}.}

\item{dist}{Character to denote to sample either from a lognormal distribution or Pareto distribution.}

\item{mu}{The mean of the distribution (default = 1).}

\item{LnSD}{If Lognormal, the standard deviation.}

\item{LnAC}{If Lognormal, the autocorrelation (in log-space).}

\item{Pshape}{If Pareto, the shape parameter. See \link[EnvStats]{Pareto}. The location parameter is calculated such that the mean = 1.}

\item{nsim_plot}{The number of simulations to plot if figure is TRUE.}

\item{SR_new}{A new stock-recruit relationship (1 = Beverton-Holt, 2 = Ricker)}

\item{h_mult}{Scalar for the new steepness value (a multiple of the old steepness parameter).}

\item{y_fit}{Length two vector for the range of years of SSB and recruit pairs used to fit the SR function.}
}
\value{
Various plots using base graphics. Returns invisibly a named list, where each entry is usually a matrix with rows indexing simulation
and columns indexing year.
}
\description{
Various plots for plotting historical time series for the operating model.
}
\details{
\code{hist_bio} returns time series of spawning biomass (SBiomass), total biomass (Biomass), abundance (Number),
recruitment (Rec), and recruitment deviates (RecDev).

\code{hist_future_recruit} returns historical and future recruitment deviations.

\code{hist_bio_schedule} plots in biological at age parameters. The corresponding array is indexed by simulation, age, and year.

\code{hist_bio_change} plots alternative projection dynamics by changing either the mean or slope. Returns an updated array of
parameters.

\code{hist_growth_I} plots histograms of von Bertalanffy parameters.

\code{hist_growth_II} plots histograms of von Bertalanffy parameters by year.

\code{hist_spatial} plots histograms of the parameters for spatial movement in a two-area model (set all to 0.5 to functionally create a single area model).

\code{hist_sel} plots selectivity/retention at age for two different years in the OM. Returns an array indexed by simulation,
age, year.

\code{hist_YieldCurve} plots the yield curve as a function of F, SPR (spawning potential ratio), spawning biomass (SBiomass),
and spawning depletion (SB_SB0). Matrices are indexed by simulation (rows) and F (columns).

\code{hist_resample_recruit} generates new recruitment deviations (with mean = 1) for the projection period. Returns
a list with an updated matrix for \code{OM@cpars$Perr_y}.

\code{hist_SRR_change} re-fits stock recruit function and generates a list with new stock-recruitment parameters \code{OM@SRrel}, \code{OM@cpars$R0}, \code{OM@cpars$hs},
and historical recruitment deviations \code{OM@cpars$Perr_y}.

\code{hist_phi0} returns a list containing a matrix (by simulation and year)
of unfished spawning biomass per recruit (\code{phi0}).

\code{hist_per_recruit} returns a list containing a matrix (by simulation and F)
of yield per recruit (YPR) and spawning potential ratio (SPR).
}
\examples{
Hist <- MSEtool::runMSE(Hist = TRUE)
bio <- hist_bio(Hist)
bio$SBiomass

# Example of backend use of `hist_SRR_change`
OM <- SubCpars(RPC::DFO_4X5Y_Haddock_2015, 1:2)
Hist <- runMSE(OM, Hist = TRUE)
vars <- hist_SRR_change(Hist, h_mult = 0.6) # Steepness is sixty percent of whatever is in the OM

OM@cpars$R0 <- vars$R0
OM@cpars$hs <- vars$h
OM@SRrel <- vars$SRrel
OM@cpars$Perr_y <- Hist@SampPars$Stock$Perr_y
OM@cpars$Perr_y[, 1:(OM@maxage + OM@nyears)] <- vars$Perr_y
Hist_new <- runMSE(OM, Hist = TRUE)
}
\author{
Q. Huynh
}
