% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LRP.R
\name{plot_LRP}
\alias{plot_LRP}
\alias{LRP_SSBhist}
\alias{LRP_SSBMSY}
\alias{LRP_SSB0}
\alias{LRP_SP}
\alias{LRP_R}
\alias{LRP_RPS}
\alias{LRP_SPR}
\alias{LRP_FMSY}
\alias{LRP_Fmed}
\alias{LRP_50Rmax}
\alias{LRP_RPS90}
\title{Plot historical dynamics relative to candidate limit reference points}
\usage{
LRP_SSBhist(
  x,
  figure = c("ts", "prob", "none"),
  SSB_y,
  prob_ratio = 1,
  prob_ylim = c(0, 1)
)

LRP_SSBMSY(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 1,
  prob_ylim = c(0, 1)
)

LRP_SSB0(
  x,
  figure = c("ts", "prob", "none"),
  type = c("equilibrium", "initial", "dynamic"),
  prob_ratio = 0.4,
  prob_ylim = c(0, 1)
)

LRP_SP(x, figure = c("ts", "phase", "none"), Bunit = c("B", "VB", "SSB"))

LRP_R(
  x,
  figure = c("ts", "SR", "none"),
  SR_xlim,
  SR_ylim,
  SR_y_RPS0,
  SR_include = 1:3
)

LRP_RPS(x, figure = c("ts", "none"))

LRP_SPR(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 0.4,
  prob_ylim = c(0, 1)
)

LRP_FMSY(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 1,
  prob_ylim = c(0, 1)
)

LRP_Fmed(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 1,
  prob_ylim = c(0, 1)
)

LRP_50Rmax(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 1,
  prob_ylim = c(0, 1),
  sims = 1:25
)

LRP_RPS90(
  x,
  figure = c("ts", "prob", "none"),
  prob_ratio = 1,
  prob_ylim = c(0, 1),
  sims = 1:25
)
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} containing
the Hist object.}

\item{figure}{Character, whether to return a time series plot (\code{"ts"}), probability plot (\code{"prob"}), or no figure (\code{"none"}).}

\item{SSB_y}{The year (relative to OM@CurentYr) to compare annual SSB values (only if prob_ratio is a numeric).}

\item{prob_ratio}{Numeric. If \code{figure = "prob"}, numeric that indicates a threshold. Functions return
annual probabilities of exceeding this threshold.}

\item{prob_ylim}{The y-axis range of the figure if \code{figure = "prob"}.}

\item{type}{For \code{LRP_SSB0}, whether the SSB0 is the equilibrium, initial, or dynamic value (see App for description).}

\item{Bunit}{The metric for stock biomass, either total biomass, vulnerable biomass, or spawning biomass.}

\item{SR_xlim}{Optional x-axis range for the stock recruit plot. Only used if \code{figure = "SR"}.}

\item{SR_ylim}{Optional y-axis range for the stock recruit plot. Only used if \code{figure = "SR"}.}

\item{SR_y_RPS0}{The year (relative to OM@CurrentYr) for which to plot unfished recruits per spawner,
Only used if \code{figure = "SR"}, and \code{any(SR_include == 3)}.}

\item{SR_include}{A vector including any of \code{c(1, 2, 3)} that indicates what to plot in the stock-recruit figure. 1 = individual S-R pairs,
2 = stock-recruit relationship, 3 = reference recruits-per-spawner (R/S) lines
(maximum R/S corresponding to stock-recruit alpha, median historical R/S, and year-specific unfished R/S).}

\item{sims}{A subset of simulations to plot (to reduce time to produce figure)}
}
\value{
Returns invisibly a list with matrices and arrays used to generate various plots associated with
time series and probabilities.
}
\description{
Generate LRP figures. See details below on various functions.
}
\details{
\code{LRP_SSBhist} returns annual spawning biomass (SSB) relative to the annual probability that SSB exceeds some historical value
(corresponding to the year in \code{SSB_y}).

\code{LRP_SSBMSY} returns annual SSB/SSBMSY or the annual probability that SSB/SSBMSY exceeds \code{prob_ratio}.

\code{LRP_SSB0} returns annual SSB/SSB0 or the annual probability that SSB/SSB0 exceeds \code{prob_ratio}.

\code{LRP_SP} returns annual surplus production (annual change in biomass - catch) and per capita surplus production.

\code{LRP_R} returns either annual recruitment (as a figure or table) or a stock recruit figure.

\code{LRP_RPS} returns annual recruits-per-spawner.

\code{LRP_SPR} returns annual spawning potential ratio.

\code{LRP_FMSY} returns annual F and F/FMSY.

\code{LRP_Fmed} returns F and Fmed (the fishing mortality corresponding to the historical median recruits per spawner,
frequently also referred to as the F-replacement F).

\code{LRP_50Rmax} returns  \code{SSB50\%Rmax}, the SSB corresponding to 50% of maximum recruitment from the stock-recruit relationship.

\code{LRP_RPS90} returns \code{SSB 90\%ile R/S}, the SSB corresponding to the intersection of the 90the percentile of both historical recruitment
and recruits-per-spawner.
}
\examples{
library(MSEtool)
Hist <- MSEtool::runMSE(MSEtool::testOM, Hist = TRUE, silent = TRUE)

LRP <- LRP_SSBhist(Hist, SSB_y = 50, prob_ratio = 0.5)
str(LRP)

LRP_50Rmax(Hist)
}
\references{
Mace, P.M. 1994. Relationships between Common Biological Reference Points Used as Thresholds and Targets of Fisheries Management Strategies.
CJFAS. 51:110-122. https://doi.org/10.1139/f94-013

Myers, et al. 1994. In search of thresholds for recruitment overfishing. ICES JMS. 51:191–205. https://doi.org/10.1006/jmsc.1994.1020
}
\author{
Q. Huynh
}
