% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc-package.R
\docType{package}
\name{rpc-package}
\alias{rpc}
\alias{rpc-package}
\title{rpc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://laijasmine.github.io/rpc/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
