% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{RPCenv}
\alias{RPCenv}
\alias{PMenv}
\alias{MPdesc}
\alias{MPinterval}
\title{RPC environments}
\format{
An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.
}
\usage{
RPCenv

PMenv

MPdesc

MPinterval
}
\description{
Environments that the shiny app will use to store information.
}
\section{Functions}{
\itemize{
\item \code{PMenv}: Stores performance metrics.

\item \code{MPdesc}: Stores descriptions of selected MPs.

\item \code{MPinterval}: Stores MP intervals.

}}
\keyword{datasets}
