% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSE_DDM.R
\name{runMSE_DDM}
\alias{runMSE_DDM}
\title{Run a management strategy evaluation with density-dependent natural mortality}
\usage{
runMSE_DDM(
  OM,
  M0,
  M1,
  MPs = "NFref",
  HistRel = FALSE,
  silent = FALSE,
  parallel = FALSE,
  extended = FALSE,
  checkMPs = FALSE
)
}
\arguments{
\item{OM}{An operating model.}

\item{M0}{Natural mortality as the stock biomass reaches unfished levels. Vector of length \code{OM@nsim}.}

\item{M1}{Natural mortality as the stock biomass reaches zero. Vector of length \code{OM@nsim}.}

\item{MPs}{A character vector of MPs.}

\item{HistRel}{Logical, whether the MICE relationship is used for the historical population. Generally `FALSE` if conditioned from
an outside model.}

\item{silent}{Should messages be printed out to the console?}

\item{parallel}{Logical. Should the MSE be run using parallel processing? See Details for more information.}

\item{extended}{Logical. Return extended projection results?}

\item{checkMPs}{Logical. Check if the specified MPs exist and can be run on SimulatedData?}
}
\value{
A \linkS4class{MSE} object.
}
\description{
**Experimental** A wrapper function that converts an \linkS4class{OM} to \linkS4class{MOM}, adds a density-dependent M relationship
(as a function of biomass) via \link{makeRel}, runs \link{multiMSE}, and converts the output back to an \linkS4class{MSE} object.
}
\details{
Assumes that fecundity-at-age, weight-at-age, and maturity-at-age are constant in the projection.

Natural mortality (M) is a linear function of stock depletion in terms to total biomass (B) in year y
(Forrest et al. 2018):
\deqn{M_y = M_0 + (M_1 + M_0) (1 - B_y/B_0)}{M_y = M_0 + (M_1 + M_0) (1 - B_y/B_0)}
with a constraint that \eqn{M_y = M_0} if \eqn{B_y > B_0}
}
\references{
Forrest, R., Holt, K., and Kronlund, A. 2018. Performance of alternative harvest control rules for two
Pacific groundfish stocks with uncertain natural mortality: Bias, robustness and trade-offs.
Fisheries Research 206: 259–286. \doi{10.1016/j.fishres.2018.04.007}
}
