% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_class}
\alias{create_class}
\title{create_class}
\usage{
create_class(start_date, sessions = 9)
}
\arguments{
\item{start_date}{(chr) in the format of YYYY-MM-DD}

\item{sessions}{(int) number of sessions}
}
\value{
data.frame containing start_date, end_date, exclusions
}
\description{
create_class
}
