% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_class}
\alias{create_class}
\title{create_class
Given a start date create class schedule}
\usage{
create_class(start_date, sessions, biweekly = FALSE)
}
\arguments{
\item{start_date}{(chr) in the format of YYYY-MM-DD}

\item{sessions}{(int) number of sessions}

\item{biweekly}{(bool) weekly or bi-weekly classes Default = FALSE}
}
\value{
tibble containing end_date, exclusions
}
\description{
create_class
Given a start date create class schedule
}
