% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_class_schedule}
\alias{get_class_schedule}
\title{get_class_schedule}
\usage{
get_class_schedule(instructor = NULL)
}
\arguments{
\item{instructor}{(chr) name of instructor}
}
\description{
get_class_schedule
}
