% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_exclusions}
\alias{get_exclusions}
\title{get exclusion dates}
\usage{
get_exclusions(start_date, sessions, class_frequency = "weekly")
}
\arguments{
\item{start_date}{(date) start date}

\item{sessions}{(integer) number of sessions}

\item{class_frequency}{(str) frequency of classes, Options = "weekly", "biweekly", Default = "weekly"}
}
\description{
get exclusion dates
}
