% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot-Hist}
\alias{plot-Hist}
\alias{hist_bio}
\alias{hist_future_recruit}
\alias{hist_bio_schedule}
\alias{hist_bio_change}
\alias{hist_growth_I}
\alias{hist_growth_II}
\alias{hist_spatial}
\alias{hist_sel}
\alias{hist_YieldCurve}
\title{Plot historical dynamics}
\usage{
hist_bio(x)

hist_future_recruit(x)

hist_bio_schedule(
  x,
  var = c("Len_age", "Wt_age", "Mat_age", "M_ageArray"),
  n_age_plot,
  yr_plot,
  sim
)

hist_bio_change(
  x,
  var = c("Wt_age", "M_ageArray"),
  change_mean = 0,
  change_slope = 0,
  n_age_plot = 10,
  sim = 1,
  figure = TRUE
)

hist_growth_I(x)

hist_growth_II(x)

hist_spatial(x, type = c("par", "matrix", "all"), ...)

hist_sel(x, yr, maturity = TRUE)

hist_YieldCurve(x, yr_bio, yr_sel, F_range)
}
\arguments{
\item{x}{An object of class \linkS4class{Hist}, or a shiny \code{reactivevalues} object containing a slot named \code{MSEhist} which is
the Hist object.}

\item{var}{A string to indicate which object to plot from OM@cpars.}

\item{n_age_plot}{The number of ages to plot in the left figure.}

\item{yr_plot}{The year (relative to OM@CurrentYr) to plot for the right figure.}

\item{sim}{The simulation to plot for the left figure.}

\item{change_mean}{The percent change in the mean of the new parameters relative to the old.}

\item{change_slope}{The percent change year-over-year in the projection parameters relative to the last projection year.}

\item{figure}{Logical, whether to return a figure (TRUE) or the updated array (FALSE) for \code{hist_bio_change}.}

\item{type}{Type of spatial plot}

\item{...}{arguments to \link[MSEtool]{plot_mov}}

\item{yr}{A length-2 vector for the years (relative to OM@CurrentYr) to plot selectivity.}

\item{maturity}{Logical, whether to plot maturity along with selectivity.}

\item{yr_bio}{The year (relative to OM@CurrentYr) for the biological parameters (growth, M, maturity).}

\item{yr_sel}{The year (relative to OM@CurrentYr) for the selectivity parameters.}

\item{F_range}{Length two vector for the range of F to plot the yield curve. By default, \code{c(1e-8, 3 * max(M))}.}
}
\value{
Various plots using base graphics
}
\description{
Various plots for plotting historical time series for the operating model.
}
\details{
\code{hist_bio} plots time series of biomass, abundance, and recruitment.

\code{hist_future_recruit} plots historical and future recruitment deviations.

\code{hist_bio_schedule} plots in biological at age parameters.

\code{hist_bio_change} plots alternative projection dynamics by changing either the mean or slope.

\code{hist_growth_I} plots histograms of von Bertalanffy parameters.

\code{hist_growth_II} plots histograms of von Bertalanffy parameters by year.

\code{hist_spatial} plots histograms of the parameters for spatial movement in a two-area model (set all to 0.5 to functionally create a single area model).

\code{hist_sel} plots selectivity/retention at age for two different years in the OM.

\code{hist_YieldCurve} plots the yield curve.
}
\examples{
Hist <- MSEtool::runMSE(Hist = TRUE)
hist_bio(Hist)
}
\author{
Q. Huynh
}
